transfer RepeatTimer into a separate internal class
